+++ peer-assessment.md
@@ -1,0 +1,61 @@
+# Self Assessment
+
+## Automate
+
+#### Running the entire analysis requires rendering one Quarto document
+
+Meets spec. All required code is in one quarto doc and accesses information from different folders and scripts adeuqately.
+
+#### The analysis runs without errors
+
+Meets spec. The analysis runs without errors. yay!!!
+
+#### The analysis produces the expected output
+
+Meets spec. The analysis produces one plot for all five categories of nutrients.
+
+#### Data import/cleaning is handled in its own script(s)
+
+Meets spec. An R data cleaning script is stored in the analysis folder.
+
+## Organize
+
+#### Raw data is contained in its own folder
+
+Meets spec. Raw data is housed within its own folder data.
+
+#### Intermediate outputs are created and saved to a separate folder from raw data
+
+Meets spec. The cleaned data is stored in output in a file full_data.
+
+#### At least one piece of functionality has been refactored into a function in its own file
+
+Meets spec. The moving average function is saved in its own script with a roxygen skeleton explaining all of the parameters.
+
+## Document
+
+#### The repo has a README that explains where to find (1) data, (2) analysis script, (3) supporting code, and (4) outputs
+
+Meets spec. The readme clearly shows the outputs and briefly explains what each link holds. 
+
+#### The README includes a flowchart and text explaining how the analysis works
+
+Not yet. No flow chart is present but a brief explanation of how the analysis works is provided. 
+
+#### The code is appropriately commented
+
+Not yet. Code is commented but not thoroughly. Some code chunks can benefit from additional comments. An example of where comments would be beneficial is in the data cleaning where data is be analyzed. Currently, someone not familiar with the function would have to dissect the code to understand the analysis. 
+
+#### Variable and function names are descriptive and follow a consistent naming convention
+
+Meets spec. Names for variables and functions are consistent and indicate what they are signifying.
+
+## Scale
+
+#### Running the environment initialization script installs all required packages
+
+Not yet. No file was found in the folder containing installation of necessary packages .
+
+#### The analysis script runs without errors
+
+Not yet. The analysis will not run on workbench yet without adequately fulfilling the above spec.
